modify code

+++ presentation/src/main/java/com/cellodove/presentation/ui/search/SearchFragment.kt
@@ -1,23 +1,19 @@
 package com.cellodove.presentation.ui.search
 
+import android.content.Context
 import android.os.Bundle
-import android.view.KeyEvent
 import android.view.View
 import android.view.inputmethod.EditorInfo
-import android.widget.Toast
+import android.view.inputmethod.InputMethodManager
 import androidx.core.view.isVisible
+import androidx.core.widget.doOnTextChanged
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
-import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
-import androidx.paging.CombinedLoadStates
-import androidx.paging.LoadState
-import androidx.paging.PagingData
-import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.DividerItemDecoration
 import com.cellodove.book.R
 import com.cellodove.book.databinding.FragmentSearchBinding
-import com.cellodove.domain.model.response.Documents
+import com.cellodove.domain.model.Documents
 import com.cellodove.presentation.base.BaseFragment
 import com.cellodove.presentation.ui.main.MainViewModel
 import com.cellodove.presentation.util.decideOnState
@@ -25,8 +21,6 @@ import com.cellodove.presentation.util.initToolbar
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.flow.distinctUntilChangedBy
-import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
@@ -44,26 +38,49 @@ class SearchFragment : BaseFragment<FragmentSearchBinding>(FragmentSearchBinding
     override fun onResume() {
         super.onResume()
         searchAdapter.setCheckLike(viewModel.getCheckLikeList())
-        searchAdapter.refresh()
     }
 
     override fun observeViewModel() {
-        refreshList()
+        statusList()
     }
 
     private fun viewInit(){
-        binding.searchRecycler.adapter = searchAdapter
         val decoration = DividerItemDecoration(requireContext(), DividerItemDecoration.VERTICAL)
+        binding.searchRecycler.adapter = searchAdapter
         binding.searchRecycler.addItemDecoration(decoration)
 
         binding.etQuery.setOnEditorActionListener { _, actionId, _ ->
             if (actionId == EditorInfo.IME_ACTION_SEARCH) {
-                viewModel.deleteAllLikeList()
-                searchBook(binding.etQuery.text.toString())
-                searchAdapter.setSearchWord(binding.etQuery.text.toString())
+                if (binding.etQuery.text.toString().isEmpty()){
+                    binding.inputLayout.error = "텍스트를 입력해 주세요."
+                }else{
+                    hideKeyboard()
+                    binding.inputLayout.error = null
+                    viewModel.deleteAllLikeList()
+                    searchBook(binding.etQuery.text.toString())
+                    binding.searchRecycler.scrollToPosition(0)
+                    searchAdapter.setSearchWord(binding.etQuery.text.toString())
+                }
                 true
             } else {
                 false
+            }
+        }
+
+        binding.etQuery.doOnTextChanged { _, _, _, _ ->
+            binding.inputLayout.error = null
+        }
+
+        binding.inputLayout.setEndIconOnClickListener {
+            if (binding.etQuery.text.toString().isEmpty()){
+                binding.inputLayout.error = "텍스트를 입력해 주세요."
+            }else{
+                hideKeyboard()
+                binding.inputLayout.error = null
+                viewModel.deleteAllLikeList()
+                searchBook(binding.etQuery.text.toString())
+                binding.searchRecycler.scrollToPosition(0)
+                searchAdapter.setSearchWord(binding.etQuery.text.toString())
             }
         }
 
@@ -79,22 +96,13 @@ class SearchFragment : BaseFragment<FragmentSearchBinding>(FragmentSearchBinding
             }
         })
 
-        searchAdapter.setLikeButtonClickListener(object  : SearchAdapter.OnLikeButtonClickListener{
-            override fun onClick(isLike: Boolean,title : String) {
-                if (isLike){
-                    viewModel.addCheckLike(title)
-                }else{
-                    viewModel.deleteCheckLike(title)
-                }
-            }
-        })
-
         binding.retryButton.setOnClickListener {
             searchAdapter.retry()
         }
     }
 
     private fun searchBook(query :String){
+        searchJob?.cancel()
         searchJob = lifecycleScope.launch {
             viewModel.searchBook(query).collectLatest {
                 binding.searchNothing.isVisible = false
@@ -106,7 +114,7 @@ class SearchFragment : BaseFragment<FragmentSearchBinding>(FragmentSearchBinding
         }
     }
 
-    private fun refreshList() {
+    private fun statusList() {
         searchAdapter.addLoadStateListener { loadState ->
             loadState.decideOnState(
                 adapter = searchAdapter,
@@ -116,16 +124,19 @@ class SearchFragment : BaseFragment<FragmentSearchBinding>(FragmentSearchBinding
                     binding.errorLayout.isVisible = false
                 },
                 showEmptyState = { visible ->
-                    binding.searchRecycler.scrollToPosition(0)
                     binding.searchNothing.isVisible = visible
                     binding.errorLayout.isVisible = false
                 },
-                showError = { message ->
+                showError = {
                     binding.searchRecycler.visibility = View.INVISIBLE
                     binding.searchNothing.isVisible = false
                     binding.errorLayout.isVisible = true
                 }
             )
         }
+    }
+
+    private fun hideKeyboard(){
+        (requireActivity().getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager).hideSoftInputFromWindow(binding.etQuery.windowToken, 0)
     }
 }
